+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## master
 
 * Fix creating a volume when server is specified by its name
+* Deprecate and ignore the `--window` flag on `hcloud server enable-backup`
 
 ## v1.9.1